+++ ansible_collections/junipernetworks/apstra/plugins/module_utils/apstra/client.py
@@ -12,6 +12,7 @@ from aos.sdk.reference_design.extension.resource_allocation import (
 )
 from aos.sdk.reference_design.extension.tags.client import Client as tagsClient
 from aos.sdk.graph import Graph
+from aos.sdk.graph import query
 import os
 import re
 import time
@@ -147,11 +148,11 @@ def _get_parent_id(parents_db, object_attrs, id):
     """
     Get the parent ids from the parent_db for an object identified by (plural) type and id.
     Ids of all parents are returned in the id dictionary.
-    
+
     :param parents_db: A dictionary of child_id to parent_id.
     :param object_attrs: A list of object types, from the root type to the last type in the id.
     :param id: The dictionary of ids.
-    
+
     :raises Exception: If the parent is not found or has no id.
     """
     # Walk backwards through the object types
@@ -178,7 +179,7 @@ _plural_to_singular = [("ies", "y"), ("s", "")]
 def plural_leaf_object_type(object_type):
     """
     Get the plural form of the leaf object type.
-    
+
     :param object_type: The object type.
     """
     attrs = object_type.split(".")
@@ -188,7 +189,7 @@ def plural_leaf_object_type(object_type):
 def singular_leaf_object_type(object_type):
     """
     Get the singular form of the leaf object type.
-    
+
     :param object_type: The object type.
     """
     plural_type = plural_leaf_object_type(object_type)
@@ -198,7 +199,7 @@ def singular_leaf_object_type(object_type):
 def singular_to_plural_id(id):
     """
     Get the plural form of the id.
-    
+
     :param id: The id dictionary.
     :return: The id dictionary with plural object types.
     """
@@ -211,7 +212,7 @@ def singular_to_plural_id(id):
 def plural_to_singular_id(id):
     """
     Get the singular form of the id.
-    
+
     :param id: The id dictionary.
     :return: The id dictionary with singular object types.
     """
@@ -224,7 +225,7 @@ def plural_to_singular_id(id):
 def singular_object_type(object_type):
     """
     Get the singular form of the object type.
-    
+
     :param object_type: The object type.
     :return: The singular form of the object type.
     """
@@ -238,7 +239,7 @@ def singular_object_type(object_type):
 def plural_object_type(object_type):
     """
     Get the plural form of the object type.
-    
+
     :param object_type: The object type.
     :return: The plural form of the object type.
     """
@@ -255,7 +256,7 @@ def plural_object_type(object_type):
 def _blueprint_lock_tag_name(blueprint_id):
     """
     Get the tag name for locking a blueprint.
-    
+
     :param blueprint_id: The ID of the blueprint.
     :return: The tag name.
     """
@@ -265,7 +266,7 @@ def _blueprint_lock_tag_name(blueprint_id):
 class ApstraClientFactory:
     """
     Factory class to create and manage Apstra clients.
-    
+
     :param module: The Ansible module.
     :param api_url: The URL of the AOS API.
     :param verify_certificates: Whether to verify SSL certificates.
@@ -274,6 +275,7 @@ class ApstraClientFactory:
     :param password: The password for authentication.
     :param logout: Whether to log out after the operation.
     """
+
     def __init__(
         self,
         module,
@@ -350,11 +352,14 @@ class ApstraClientFactory:
                 # Map the object type to the client
                 self.network_objects_set[object_type] = object_client
 
+        # Blueprint query can be cached
+        self._blueprint_graph = None
+
     @classmethod
     def from_params(cls, module):
         """
         Create an ApstraClientFactory from the module parameters.
-        
+
         :param module: The Ansible module.
         """
         api_url = module.params.get("api_url")
@@ -508,7 +513,7 @@ class ApstraClientFactory:
         """
         Call object op. If op is 'get', will get one object, or all objects of that type.
         Internal method uses the plural types to simplify logic
-        
+
         :param object_type: The object type.
         :param op: The operation to perform.
         :param id: The id dictionary.
@@ -725,7 +730,7 @@ class ApstraClientFactory:
         Lock the blueprint with the given ID.
         This is a "gentlemen's agreement" lock, not a true lock.
         A tag is used for locking.
-        
+
         :param id: The ID of the blueprint to lock.
         :param timeout: The maximum time to wait for the blueprint to be locked.
         :return: True if the blueprint was locked, False if not.
@@ -754,13 +759,21 @@ class ApstraClientFactory:
                 if re.search(locked_pattern, error_message):
                     time_left = timeout - (time.time() - start_time)
                     if time_left <= 0:
-                        self.module.fail_json(msg=f"Failed to lock blueprint {id} within {timeout} seconds")
-                    self.module.debug(f"Blueprint {id} is locked, waiting up to {time_left} seconds for unlock...")
+                        self.module.fail_json(
+                            msg=f"Failed to lock blueprint {id} within {timeout} seconds"
+                        )
+                    self.module.debug(
+                        f"Blueprint {id} is locked, waiting up to {time_left} seconds for unlock..."
+                    )
                     time.sleep(interval)
                 else:
-                    self.module.fail_json(msg=f"Unexpected ClientError trying to lock blueprint {id} within {timeout} seconds: {e}")
+                    self.module.fail_json(
+                        msg=f"Unexpected ClientError trying to lock blueprint {id} within {timeout} seconds: {e}"
+                    )
             except Exception as e:
-                self.module.fail_json(msg=f"Unexpected Exception trying to lock blueprint {id} within {timeout} seconds: {e}")
+                self.module.fail_json(
+                    msg=f"Unexpected Exception trying to lock blueprint {id} within {timeout} seconds: {e}"
+                )
 
     def unlock_blueprint(self, id):
         """
@@ -795,7 +808,7 @@ class ApstraClientFactory:
     def commit_blueprint(self, id, timeout=DEFAULT_BLUEPRINT_COMMIT_TIMEOUT):
         """
         Commit the blueprint with the given ID.
-        
+
         :param id: The ID of the blueprint to commit.
         :param timeout: The maximum time to wait for the blueprint to be committed.
         :return: True if the blueprint was committed, False if not.
@@ -808,8 +821,12 @@ class ApstraClientFactory:
             blueprint = blueprint_client.blueprints[id].get()
             time_left = timeout - (time.time() - start_time)
             if time_left <= 0:
-                self.module.fail_json(msg=f"Failed to commit blueprint {id} within {timeout} seconds")
-            self.module.debug(f"Waiting {time_left} seconds for blueprint to be committed...")
+                self.module.fail_json(
+                    msg=f"Failed to commit blueprint {id} within {timeout} seconds"
+                )
+            self.module.debug(
+                f"Waiting {time_left} seconds for blueprint to be committed..."
+            )
             time.sleep(interval)
 
         is_committed = blueprint.is_committed()
@@ -820,7 +837,7 @@ class ApstraClientFactory:
     def compare_and_update(self, current, desired, changes):
         """
         Recursively compare and update the current state to match the desired state.
-        
+
         :param current: The current state dictionary.
         :param desired: The desired state dictionary.
         :param changes: A dictionary to track changes.
@@ -830,15 +847,19 @@ class ApstraClientFactory:
         for key, desired_value in desired.items():
             if key not in current:
                 # Field is missing in the current state, probably only for create
-                self.module.debug(f"Field '{key}' missing in current state, ignoring it")
+                self.module.debug(
+                    f"Field '{key}' missing in current state, ignoring it"
+                )
                 continue
-            
+
             current_value = current[key]
-            
+
             if isinstance(desired_value, dict) and isinstance(current_value, dict):
                 # Recursively compare nested dictionaries
                 nested_changes = {}
-                nested_changed = self.compare_and_update(current_value, desired_value, nested_changes)
+                nested_changed = self.compare_and_update(
+                    current_value, desired_value, nested_changes
+                )
                 if nested_changed:
                     changes[key] = nested_changes
                     changed = True
@@ -853,7 +874,7 @@ class ApstraClientFactory:
                 current[key] = desired_value
                 changes[key] = desired_value
                 changed = True
-        
+
         return changed
 
     def extract_field(self, data, field_name):
@@ -868,12 +889,25 @@ class ApstraClientFactory:
         if field_name not in data:
             # Nothing to do
             return None
-        
+
         # Extract the field
         field_value = data.pop(field_name)
-        
+
         # Return the field in a new dictionary
         return {field_name: field_value}
+
+    def get_blueprint_graph(self, blueprint_id):
+        """
+        Get the blueprint with the given ID.
+
+        :param blueprint_id: The ID of the blueprint to get.
+        :return: The blueprint.
+        """
+        if not self._blueprint_graph:
+            blueprint_client = self.get_client("blueprints")
+            self._blueprint_graph = blueprint_client.blueprints[blueprint_id].get()
+
+        return self._blueprint_graph
 
     def update_tags(self, id, leaf_type, tags):
         """
@@ -891,20 +925,83 @@ class ApstraClientFactory:
         # Get the set of tags for the blueprint
         tags_client = self.get_tags_client()
         all_tags = tags_client.blueprints[blueprint_id].tags.list()
-        all_tags_set = {tag['label'] for tag in all_tags if 'label' in tag}
-        
+        all_tags_set = {tag["label"] for tag in all_tags if "label" in tag}
+
         # Create a set of requested tags for quick lookup
         tags_set = set(tags)
 
         # Make sure all requested tags are present
         if not all_tags_set.issuperset(tags_set):
             missing_tags = tags_set.difference(all_tags_set)
-            self.module.fail_json(msg=f"update_tags failed: missing tags: {missing_tags}")
+            self.module.fail_json(
+                msg=f"update_tags failed: missing tags: {missing_tags}"
+            )
 
         # Find out what tags are not set
         missing_tags = all_tags_set.difference(tags_set)
 
         # Update the tags
-        tags_client.blueprints[blueprint_id].tagging([id[leaf_type]], tags, list(missing_tags))
-        self.module.debug(f"Tags updated for {leaf_type} {id}, ADDED: {tags}, REMOVED: {missing_tags}")
+        tags_client.blueprints[blueprint_id].tagging(
+            [id[leaf_type]], tags, list(missing_tags)
+        )
+        self.module.debug(
+            f"Tags updated for {leaf_type} {id}, ADDED: {tags}, REMOVED: {missing_tags}"
+        )
         return tags
+
+    def query_blueprint(self, blueprint_id, graph_query):
+        """
+        Query the blueprint with the given ID.
+
+        :param blueprint_id: The ID of the blueprint to query.
+        :param query_string: The query string.
+        :return: The result of the query.
+        """
+        bp = self.get_blueprint_graph(blueprint_id)
+
+        result = list(query.iterate(bp, graph_query))
+        return result
+
+    def get_by_label(self, blueprint_id, obj_type, label, label_key="label"):
+        """
+        Find an object by label.
+
+        :param blueprint_id: The ID of the blueprint to query.
+        :param obj_type: The query string.
+        :param label: The label of the object to find.
+        :param label_key: The key to use for the label.
+        :return: The result of the query.
+        """
+        bp = self.get_blueprint_graph(blueprint_id)
+
+        get_query = query.node(type=obj_type, **{label_key: label}, name=obj_type)
+
+        result = list(query.iterate(bp, get_query))
+        if not result:
+            self.module.debug(f"Object with {label_key} {label} not found")
+            return result
+
+        if len(result) > 1:
+            self.module.fail_json(
+                msg=f"Multiple objects with {label_key} {label} found"
+            )
+
+        if obj_type not in result[0]:
+            self.module.fail_json(msg=f"Object missing key {obj_type}: {result[0]}")
+
+        return result[0][obj_type]
+
+    def get_id_by_label(self, blueprint_id, obj_type, label, label_key="label"):
+        """
+        Find an object by label and return its ID.
+
+        :param blueprint_id: The ID of the blueprint to query.
+        :param obj_type: The query string.
+        :param label: The label of the object to find.
+        :param label_key: The key to use for the label.
+        :return: The ID of the object.
+        """
+        obj = self.get_by_label(blueprint_id, obj_type, label, label_key)
+        if not obj:
+            return None
+        return obj.id