+++ ansible_collections/junipernetworks/apstra/docs/CHANGELOG.rst
@@ -4,6 +4,28 @@ Junipernetworks Apstra Collection Release Notes
 
 .. contents:: Topics
 
+v0.1.15
+=======
+
+Major Changes
+-------------
+
+- Update application points by label instead of ID.
+
+Minor Changes
+-------------
+
+- Find objects by label with the graph API.
+- Look up endpoint policies by virtual network label.
+
+v0.1.14
+=======
+
+Minor Changes
+-------------
+
+- Replace node_type parameter in apstra_facts with more generic filter parameter. Default behavior is unchanged for nodes.
+
 v0.1.13
 =======